Now dealing with seconds as doubles
- Why? The main reason was I'm thinking this will be more cross platform, as if the specified time type is an unsigned type and you try to get the difference between two times where the result should be negative, since it's unsigned you will actually get something huge. As a double, the user can provide a function to convert the raw time value into seconds (as a double) and then we can easily get the difference of two times.
- Previous commit's tests now pass.

+++ source/include/registration_api.h
@@ -35,7 +35,7 @@ void IoTLib_register_retrieve_last_upload_time_function(IoTLib_time_t (*retireve
 void IoTLib_register_retrieve_all_stored_unsent_sensor_data_function(struct IoTLib_RawSensorDataAndSensorID* (*retrieveAllUnsentDataFunc)());
 void IoTLib_register_get_stored_unsent_data_count_function(size_t (*getStoredUnsentDataCountFunc)());
 void IoTLib_register_get_current_time_function(IoTLib_time_t (*getCurrentTimeFunc)());
-void IoTLib_register_convert_time_type_to_seconds_func(uint32_t (*convertTimeTypeToSecondsFunc)(IoTLib_time_t rawTime));
+void IoTLib_register_convert_time_type_to_seconds_func(double (*convertTimeTypeToSecondsFunc)(IoTLib_time_t rawTime));
 
 #ifdef __cplusplus
 }
